Moved the public interface functions into the entry point module
